Restore coverage from the examples
The examples seemed to be running against a different version of
Incremental (possibly from pip's cache?). Ensure they run against
exactly the version of Incremental we're trying to test against by
pre-installing it in an isolated environment.

Since this introduces an isolated environment, also install coverage-p
to enable coverage of the subprocess.